requirements.txt
phidata
streamlit
openai
duckduckgo-search
requests
python-dotenv
packaging
